[revert] Change MatMul before Softmax back to @ operator
Reverts commit 9bda0a85dc5848b2c228cc1cf85fbeb67311db6b

This should be advantageous because we omit on of the two re-quantization steps between MatMul and Softmax. With this approach the resolution of the Softmax inputs should be better. At the same time it restore compatibility with I-BERT type of Softmax which should directly work on 32bit inputs.